Move ccpo routes into their own file
