+++ templates/ccpo/confirm_user.html
@@ -6,7 +6,7 @@
 {% block content %}
   {% if new_user %}
     {% call Alert('Confirm new CCPO user') %}
-      <form id="add-ccpo-user-form" action="{{ url_for('atst.confirm_new_ccpo_user') }}" method="POST">
+      <form id="add-ccpo-user-form" action="{{ url_for('ccpo.confirm_new_ccpo_user') }}" method="POST">
         {{ form.csrf_token }}
         <input type="hidden" name="dod_id" value="{{ form.dod_id.data }}">
         <div>
@@ -26,7 +26,7 @@
               v-bind:disabled="invalid"
               class='action-group__action usa-button'
               value='Confirm and Add User'>
-          <a class='action-group__action icon-link icon-link--default' href="{{ url_for('atst.ccpo_users') }}">{{ "common.cancel" | translate }}</a>
+          <a class='action-group__action icon-link icon-link--default' href="{{ url_for('ccpo.ccpo_users') }}">{{ "common.cancel" | translate }}</a>
         </div>
       </form>
     {% endcall %}
@@ -34,7 +34,7 @@
     {% call Alert('User not found') %}
       To add someone as a CCPO user, they must already have an ATAT account.
       <div class='action-group'>
-        <a class='action-group__action icon-link icon-link--default' href="{{ url_for('atst.ccpo_users') }}">
+        <a class='action-group__action icon-link icon-link--default' href="{{ url_for('ccpo.ccpo_users') }}">
           Return to list of CCPO users
         </a>
       </div>
