+++ templates/ccpo/users.html
@@ -31,7 +31,7 @@
   </div>
 
   {% if user_can(permissions.CREATE_CCPO_USER) %}
-    <a class="icon-link modal-link" href="{{ url_for('atst.add_new_ccpo_user')}}">
+    <a class="icon-link modal-link" href="{{ url_for('ccpo.add_new_ccpo_user')}}">
       Add new CCPO user {{ Icon("plus") }}
     </a>
   {% endif %}
